+++ VKTS/include/vkts/core/binary_buffer/fn_binary_buffer.hpp
@@ -36,13 +36,13 @@ namespace vkts
  *
  * @ThreadSafe
  */
-VKTS_APICALL IBinaryBufferSP VKTS_APIENTRY binaryBufferCreate(const size_t size);
+VKTS_APICALL IBinaryBufferSP VKTS_APIENTRY binaryBufferCreate(const uint32_t size);
 
 /**
  *
  * @ThreadSafe
  */
-VKTS_APICALL IBinaryBufferSP VKTS_APIENTRY binaryBufferCreate(const uint8_t* data, const size_t size);
+VKTS_APICALL IBinaryBufferSP VKTS_APIENTRY binaryBufferCreate(const uint8_t* data, const uint32_t size);
 
 }
