+++ VKTS/include/vkts/core/json/JSONobject.hpp
@@ -55,7 +55,7 @@ public:
 
 	const Vector<std::string>& getAllKeys() const;
 
-	size_t size() const;
+	uint32_t size() const;
 
 	virtual VkBool32 encode(std::string& jsonText, std::int32_t& spaces) const;
